* Added fluent styles for CheckBox, RadioButton and AccentButton * Control accent color now depends on app theme * Refactored border brush of some controls * Refactored background and foreground of some controls * Fixed parent path root folder in file path * Added confirmation dialog * Fixed button margins in content dialog * Delete now requires confirmation * Fixed paste context menu plurality * Fixed delete file being performed twice * Fixed style source of force fluent styles restart warning icon

+++ ADB Explorer/MainWindow.xaml.cs
@@ -2079,8 +2079,17 @@ namespace ADB_Explorer
             DeleteFiles();
         }
 
-        private void DeleteFiles()
+        private async void DeleteFiles()
         {
+            var result = await DialogService.ShowConfirmation(
+                $"Are you sure you want to delete {(selectedFiles.Count() == 1 ? DisplayName(selectedFiles.First()) : selectedFiles.Count() + " files")}?",
+                "Confirm Delete",
+                "Delete",
+                icon: DialogService.DialogIcon.Delete);
+
+            if (result is not ContentDialogResult.Primary)
+                return;
+
             try
             {
                 ShellFileOperation.DeleteItems(CurrentADBDevice, selectedFiles);
@@ -2469,7 +2478,10 @@ namespace ADB_Explorer
         private void ExplorerGrid_PreviewKeyDown(object sender, KeyEventArgs e)
         {
             if (e.Key == Key.Delete)
+            {
+                e.Handled = true;
                 DeleteFiles();
+            }
         }
     }
 }
